# openapi.yaml
openapi: 3.0.3
info:
  title: Bug Hunting Record & Analysis API
  version: 1.0.0
  description: >
    API for recording HTTP requests, managing notes, attachments, programs, and endpoints,
    and analyzing them for bug hunting.
    Errors are always returned as `text/plain` with a string message.
    Authentication is handled by starting a session with `/start_session`.
servers:
  - url: http://localhost:8081

# === Authentication ===
paths:
...
  /attachments:
    post:
      summary: Upload an attachment
      description: Upload a file linked to a program, endpoint, or request
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [reference_type, reference_id, file] # reference_type and reference_id are hidden and will comes from the associated Endpoint, Program or Request
              properties:
                reference_type:
                  type: string
                  enum: [programs, endpoints, requests]
                  description: Type of resource the attachment is associated with
                  example: "endpoints"
                reference_id:
                  type: integer
                  description: ID of the referenced resource
                  example: 123
                file:
                  type: string
                  format: binary
                  description: File to upload
      responses:
        "201":
          description: Attachment created successfully (returns the attachment ID as plain text)
          content:
            text/plain:
              schema:
                type: integer
                example: 555
        "400":
          description: Invalid input
          content:
            text/plain:
              schema:
                type: string
                example: "invalid file"

    delete:
      summary: Delete an attachment
      description: Deletes an attachment by ID
      parameters:
        - name: id
          in: query
          description: ID of the attachment to delete
          required: true
          schema:
            type: integer
            example: 555
      responses:
        "204":
          description: Attachment deleted successfully
        "404":
          description: Attachment not found
          content:
            text/plain:
              schema:
                type: string
                example: "attachment not found"

# === Programs ===
...